separate file for token manager
